% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats133.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{The number of trials}

\item{p}{The probability in 1 trial}
}
\value{
A data frame with two classes, bindis and data.frame
}
\description{
A data frame of probility by success
}
\examples{
bin_distribution(n = 5, p = 0.5)
}
