% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats133.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(n, p)
}
\arguments{
\item{n}{The number of trials}

\item{p}{The probability in 1 trial}
}
\value{
A number representing the "tailedness"
}
\description{
A measure of the "tailedness" of the probability distribution of a random variable.
}
\examples{
bin_kurtosis(10, 0.3)
}
