% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats133.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(n, p)
}
\arguments{
\item{n}{The number of trials}

\item{p}{The probability in 1 trial}
}
\value{
The expected number of successes in n trials
}
\description{
Calculates the expected number of successes in n trials
}
\examples{
bin_mean(10, 0.03)
}
