% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats133.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(n, p)
}
\arguments{
\item{n}{The number of trials}

\item{p}{The probability in 1 trial}
}
\value{
the most likely number of success in n independent trials w probility of success on each trial
}
\description{
Calculates the most likely number of success in n independent trials w probility of success on each trial
}
\examples{
bin_mode(10, 0.03)
}
