% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats133.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(k, n, p)
}
\arguments{
\item{k}{Possible successes}

\item{n}{The number of trials}

\item{p}{The probability of success in 1 trial}
}
\value{
The probability
}
\description{
Calculates the probability of getting x success in y trials
}
\examples{
bin_probability(k = 2, n = 5, p = 0.5)
bin_probability(k = 0:2, n = 5, p = 0.5)
bin_probability(k = 55, n = 100, p = 0.45)
}
