% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats133.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(n, p)
}
\arguments{
\item{n}{The number of trials}

\item{p}{The probability in 1 trial}
}
\value{
A number representing the variance
}
\description{
a measure of asymmetry of the probability distribution of a random variable about its mean.
}
\examples{
bin_skewness(10, 0.03)
}
