% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats133.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{The number of trials}

\item{p}{The probability in 1 trial}
}
\value{
A list with named elements trials and prob
}
\description{
A list of class binvar
}
\examples{
bin_variable(n = 5, p = 0.5)
}
