% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats133.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(n, p)
}
\arguments{
\item{n}{The number of trials}

\item{p}{The probability in 1 trial}
}
\value{
A number representing the variance
}
\description{
The variance using the auxiliary function
}
\examples{
bin_variance(10, 0.03)
}
