% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_choose}
\alias{bin_choose}
\title{function called bin choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{is the number trials}

\item{k}{is the number of successes}
}
\value{
is the computed number of combinations
}
\description{
computes the number of combinations in which k success can occur in n trials
}
\examples{
#choose a success of 2 out of 10 trials

bin_choose(n=5,k=2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
