% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{function used to compute the bin distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the bin distribution in form of table and a histogram
}
\description{
computes the distribution of the bin probability
}
\examples{
#default bin distribution

bin_distribution(trials=5, prob=0.5)
dis1 <- bin_distribution(trials=5, prob = 0.5)
plot(dis1)
}
