% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_mean}
\alias{bin_mean}
\title{compute the binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the binomial mean of an object
#' @examples

#bin_mean
bin_mean(10,0.2)
}
\description{
generates binomial mean
}
