% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{function used to compute bin variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the binomial random variable object
}
\description{
generates bin variable
}
\examples{

#bin variable example
bin1 <- bin_variable(trials=10, p=0.3)
bin1
bin1 <- bin_variable(trials = 10, p =0.3)
binsum1 <- summary(bin1)
binsum1
}
