% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variance}
\alias{bin_variance}
\title{compute the binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{is the probability of success}
}
\value{
returns the binomial variance of an object
#' @examples

#bin_variance
bin_variance(10,0.2)
}
\description{
generates binomial variance
}
